Renamed class Parameters and added it to the arguments of ADAP3DFeatureDetectionMethod
