+++ msdk-featuredetection-adap3d/src/main/java/io/github/msdk/featdet/ADAP3D/common/algorithms/BiGaussian.java
@@ -52,7 +52,7 @@ public class BiGaussian {
 
   /**
    * <p>
-   * Inside BiGaussian Constructor we're determining 4 BiGaussian Parameters. MaxHeight, Mu,
+   * Inside BiGaussian Constructor we're determining 4 BiGaussian ADAP3DFeatureDetectionParameters. MaxHeight, Mu,
    * SigmaLeft and SigmaRight.
    * </p>
    *
